répartition javascript et suppression print

+++ app/models.py
@@ -391,7 +391,6 @@ def next_lieu_id():
 def add_lieu_stock(nom_lieu):
     existing_lieu = Lieu_Stockage.query.filter_by(nomLieu=nom_lieu).first()
     if not existing_lieu:
-        print("kk")
         id = next_lieu_id()
         lieu = Lieu_Stockage(id, nom_lieu)
         db.session.add(lieu)
@@ -439,7 +438,6 @@ def get_all_chimiste():
 
 def get_pagination_produits(page=1, nb=15):
     # Pour les produits non caché
-    print("testergdfgdfg")
     liste_prod_qte = []
     liste_prod = Produit.query.filter(Produit.afficher == 1).all()
     liste_prod_cacher = Produit.query.filter(Produit.afficher == False).all()
@@ -1010,25 +1008,27 @@ def convertir_quantite(id_produit):
 
 def update_etat(idCommande, idChimiste):
     faire = Faire.query.filter(Faire.idCommande == idCommande).first()
-    match faire.statutCommande:
-        case 'non-commence':
-            faire.idChimiste = idChimiste
-            faire.statutCommande = 'en-cours'
-        case 'en-cours':
-            faire.statutCommande = 'termine'
-    db.session.commit()
+    if faire is not None:
+        match faire.statutCommande:
+            case 'non-commence':
+                faire.idChimiste = idChimiste
+                faire.statutCommande = 'en-cours'
+            case 'en-cours':
+                faire.statutCommande = 'termine'
+        db.session.commit()
 
 def delete_reservation(idCommande, idChimiste):
     commande = Commande.query.get(idCommande)
-    faire = Faire.query.filter(Faire.idCommande == idCommande).first()
-    print(Chimiste.query.get(idChimiste).estPreparateur, faire.statutCommande)
-    if faire.statutCommande == 'non-commence' or (Chimiste.query.get(idChimiste).estPreparateur and faire.statutCommande == "en-cours"):
-        produit = Produit.query.get(commande.idProduit)
-        est_Stocker = Est_Stocker.query.filter(Est_Stocker.idProduit == produit.idProduit).first()
-        est_Stocker.quantiteStocke += commande.qteCommande
-        db.session.delete(faire)
-        db.session.delete(commande)
-        db.session.commit()
+    if commande is not None:
+        faire = Faire.query.filter(Faire.idCommande == idCommande).first()
+        if faire is not None:
+            if faire.statutCommande == 'non-commence' or (Chimiste.query.get(idChimiste).estPreparateur and faire.statutCommande == "en-cours"):
+                produit = Produit.query.get(commande.idProduit)
+                est_Stocker = Est_Stocker.query.filter(Est_Stocker.idProduit == produit.idProduit).first()
+                est_Stocker.quantiteStocke += commande.qteCommande
+                db.session.delete(faire)
+                db.session.delete(commande)
+                db.session.commit()
 
 def est_en_dessous_du_seuil():
     """Renvoie une liste des produits qui sont en dessous de leur seuil
